# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto API
  description: |-
    This OpenAPI document describes WASAphoto functions
  version: "1"
...
security:
  - bearerAuth: []
paths:
...
  /profiles/{user_id}/photos/{photo_id}:
    parameters:
      - name: user_id
        in: path
        description: ID of user
        schema: { $ref: "#/components/schemas/UserID"}
        required: true
      - name: photo_id
        in: path
        description: ID of photo
        schema: { $ref: "#/components/schemas/PhotoID"}
        required: true
    delete:
        tags: ["action"]
        operationId: deletePhoto
        summary: user deleting a photo
        description: remove a photo from user profile
        responses:
          "204":
            description: Photo successfully deleted
          "400":
            description: Bad Request - Invalid user_id or photo_id
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: {error: "Error while retrieving user ID from parameters"}
          "401":
            description: Unauthorized - Invalid or missing token
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: {error: "Unauthorized"}
          "404":
            description: Not Found - Photo does not exist
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: {error: "Photo not found"}
          "500":
            description: Internal Server Error - Server encountered an error
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: {error: "Error while trying to delete photo from database"}
    get:
      tags:
        - action
      operationId: getPhoto
      summary: Retrieve a specific photo by its ID
      description: Get details of a photo uploaded by the user using the photo ID.
      responses:
        '200':
          description: Photo successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPhoto"
              example:
                user_id: 1
                pid: 45
                photo: "base64-encoded-image-data"
                caption: "A beautiful sunset"
                time: "2023-11-21T23:19:59"
        '400':
          description: Bad request (invalid photo ID)
        '404':
          description: Photo not found
        '500':
          description: Internal server error

    put:
      tags: ["action"]
      operationId: followUser
      summary: add new user to followed list
      description: todo
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/User"}
      responses:
        "200":
          description: user successfully added to followed list
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Followed"}
              example: {username: Marco, user_id: 128}
#    delete:
#      tags: ["action"]
#      operationId: unfollowUser
#      summary: "unfollow a user"
#      description: removing from the followed list a user
#      responses:
#       "204":
#          description: user successfully unfollowed
  /profiles/{user_id}/bans/{target_uid}:
    parameters:
      - name: user_id
        in: path
        description: user's id that wants to ban/unban another user
        schema: {$ref: "#/components/schemas/UserID"}
        required: true
      - name: target_uid
        in: path
        description: user's id that is going to be banned/unbanned
        schema: {$ref: "#/components/schemas/UserID"}
        required: true
    put:
      tags: ["action"]
      operationId: banUser
      summary: ban a user
      description: add the target_uid user to user's current_uid banned list
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/User" }
      responses:
        "200":
          description: user successfully added to ban list
          content:
            application/json:
              schema: {$ref: "#/components/schemas/User"}
        "400":
          description: Invalid request (e.g., invalid user_id or target_uid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id or target_uid"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"
    delete:
      tags: ["action"]
      operationId: unbanUser
      summary: unban a user
      description: remove the target_uid user from user's current_uid banned list
      responses:
        "204":
          description: user successfully unbanned
        "400":
          description: Invalid request (e.g., invalid user_id or target_uid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id or target_uid"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"
    get:
      tags: ["action"]
      operationId: getBan
      summary: get the user who banned another user
      responses:
        "200": 
          description: user successfully retrieved
          content:
            application/json:
              schema: {$ref: "#/components/schemas/User"}
        "400":
          description: Invalid request (e.g., invalid user_id or target_uid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id or target_uid"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"
...
  /photos/{photo_id}/likes:
    parameters:
      - name: photo_id
        in: path
        description: ID of the photo to retrieve likes for
        required: true
        schema: { $ref: "#/components/schemas/PhotoID" }

    get:
      tags: ["action"]
      summary: Retrieve likes for a photo
      description: |
        Retrieve a list of users who liked the specified photo.
      operationId: getLikes
      responses:
        "200":
          description: Successfully retrieved the list of likes
          content:
            application/json:
              schema:
                type: array
                description: The list of likes
                minItems: 0
                maxItems: 1000000
                items:
                  type: object
                  properties:
                    userID: {$ref: "#/components/schemas/UserID" }
                    username: {$ref: "#/components/schemas/Username" }
        "400":
          description: Bad request. The request could not be understood or was missing required parameters.
        "401":
          description: Unauthorized. Authentication is required and has failed or has not been provided.
        "404":
          description: Not found. The specified photo could not be found.
        "500":
          description: Internal server error. An error occurred on the server.
    post:
      tags: ["action"]
      operationId: commentPhoto
      summary: comment a photo
      description: user with current_id adds to Comments list of user with target_id    
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Comment"}
      responses:
        "200":
          description: user successfully commented post
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Comment"}
              example: {comment_id: 7, comment_text: "good panorama!"}

    delete:
      tags: ["action"]
      operationId: uncommentPhoto
      summary: remove a comment from photo
      description: user with current_id removes from user's comment list with target_id
      responses:
        "200":
          description: user successfully removed a comment from photo 
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CommentID"}
              example: {comment_id: 7}
  /profiles/{user_id}/comments/{photo_id}:
    parameters:
      - name: user_id
        in: path
        description: The user ID of the user who will receive the comment.
        schema: {  $ref: "#/components/schemas/UserID" }
        required: true
      - name: photo_id
        in: path
        description: The photo ID on which the comment is made.
        schema: { $ref: "#/components/schemas/PhotoID" }
        required: true
    
    post:
      tags:
        - action
      operationId: commentPhoto
      summary: Comment on a photo
      description: Adds a comment to the list of comments on a photo for a user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
            examples:
              example1:
                value:
                  comment_text: "good panorama!"
      responses:
        '200':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
              example:
                comment_id: 7
                comment_text: "good panorama!"
        '400':
          description: Bad request, possibly invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: integer
                    description: The HTTP error code
                  error_message:
                    type: string
                    description: A description of the error
                required:
                  - error_code
                  - error_message
              example:
                error_code: 400
                error_message: "Invalid comment data"
        '404':
          description: User or photo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: integer
                    description: The HTTP error code
                  error_message:
                    type: string
                    description: A description of the error
                required:
                  - error_code
                  - error_message
              example:
                error_code: 404
                error_message: "User or photo not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: integer
                    description: The HTTP error code
                  error_message:
                    type: string
                    description: A description of the error
                required:
                  - error_code
                  - error_message
              example:
                error_code: 500
                error_message: "Server error"
...
  /profiles/{user_id}/ban:
    parameters:
      - name: user_id
        in: path
        description: The user's ID whose ban list is to be retrieved
        schema: {$ref: "#/components/schemas/UserID"}
        required: true
    get:
      tags: ["action"]
      operationId: getBanList
      summary: Get a user's ban list
      description: Retrieve the list of users banned by the specified user
      responses:
        "200":
          description: Successfully retrieved the ban list
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Banned"}
        "400":
          description: Invalid request (e.g., invalid user_id)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"

#COMPONENTS
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommentID:
      title: Comment ID 
      description: id of a comment
      type: object
      properties:
        id:
          description: id of a comment
          type: integer
          example: 58
    CommentText:
      title: Comment Text
      description: a string representing a text under a post
      type: object
      properties:
        text:
          type: string
          description: a string representing a text under a post
          example: hey,nice photo
          pattern: '^[^\/\\]{1,64}$'
          maxLength: 100
          minLength: 1
    Comment:
      title: Comment
      description: "text added from users to posts"
      type: object
      properties:
        username:
          $ref: "#/components/schemas/Username"
        comment_id:
          $ref: "#/components/schemas/CommentID"
        comment_text:
          $ref: "#/components/schemas/CommentText"
        timestamp:
          type: string
          description: date and time when the comment is posted
          format: date-time
          example: "2010-01-20T16:10:19Z"
          minLength: 20
          maxLength: 20
          readOnly: true
    Liked:
      title: Liked list
      description: list of all liked photos
      type: object
      properties:
        photo_id:
          type: array
          description: list of all liked photos
          items: {$ref: "#/components/schemas/PhotoID"}
          minItems: 0
          maxItems: 10000000
...
    PhotoID:
      title: photo id
      description: "photo identificator"
      type: object
      properties:
        id:
          type: integer
          example: 45
    Followed:
      title: Followed list
      description: list of users' names and IDs followed by user
      type: object
      properties:
        user:
          type: array
          description: list of users' names and IDs followed by user
          items: 
            oneOf:
              - {$ref: "#/components/schemas/Username"}
              - {$ref: "#/components/schemas/UserID"}
          minItems: 0
          maxItems: 10000000
    Banned:
      title: Ban list
      description: list of users' names and IDs banned by user
      type: object
      properties:
        user:
          type: array
          description: list of users' names and IDs banned by user
          items:
            oneOf:
              - {$ref: "#/components/schemas/Username"}
              - {$ref: "#/components/schemas/UserID"}
          minItems: 0
          maxItems: 10000000
    UserPhoto:
      title: User Photo
      description: Details of an uploaded photo
      type: object
      properties:
        photo_id: {$ref: "#/components/schemas/PhotoID"}
        user_id: {$ref: "#/components/schemas/UserID"}
        image:
          $ref: '#/components/schemas/Image'
        timestamp:
          type: string
          description: date and time when the image was posted
          format: date-time
          example: "2010-09-20T16:10:19Z"
          minLength: 20
          maxLength: 20
          readOnly: true
        likeCount:
          description: number of likes that the photo received
          type: integer
          example: 10
          minimum: 0
          readOnly: true
        commentCount:
          description: number of comments that the photo received
          type: integer
          example: 10
          minimum: 0
          readOnly: true
        liked:
          description: flag that retrieves if the users liked to the photo
          type: boolean
          example: false
          readOnly: true
    # UserPhoto:
    #   title: User photo
    #   description: photo uploaded by the user
    #   properties:
    #     user_id: {$ref: "#/components/schemas/UserID"}
    #     pid: {$ref: "#/components/schemas/PhotoID"}
    #     photo: {$ref: "#/components/schemas/Image"}
#          title: photo path
#          description: "describes the path of a photo"
#          type : string
#          pattern: '^\/(?:[\w]+\/)*[\w]+\.[a-zA-Z]{2,}$'
#          example: "path/images/photo.png"
#          minLength: 1
#          maxLength: 100
          
        caption: {$ref: "#/components/schemas/Caption"}
          
#        likes_number:
#          title: likes number
#          description: "number of likes received from users"
#          type: integer
#          example: 500
#        comments_number:
#          title: comments number
#          description: "number of comments received from users"
#          type: integer
#          example: 5
#        comments_list:
#          title: comments list
#          description: array of Comment
#          type: array
#          items: {$ref: "#/components/schemas/Comment"}
#          minItems: 0
#          maxItems: 10000000
        time:
          title: time
          description: pubblication time
          type: string
          format: date-time
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
          example: "2023-11-21T23:19:59"
          minLength: 19
          maxLength: 19
    Image: 
          title: photo
          description: base 64 image.
          type: string
          format: byte
          minLength: 0
          maxLength: 1000000000
    Caption:
          description: "description of a photo"
          type: string
          pattern: '^[^\/\\]{1,64}$'
          example: "look at me!"
          minLength: 0
          maxLength: 50
...
    UserID:
      title: User ID
      description: "this is the user id"
      type: object
      properties:
        id:
          description: "user identificator number"
          type: integer
          example: 1
#          readOnly: true
    Username:
      title: Username
      description: "this is the user's name"
      type: object
      properties:
        username:
          type: string
          description: "name of the user"
          example: "Simone"
          pattern: '^[a-z][a-z0-9]*$'
          minLength: 3
          maxLength: 16
    User:
      title: User
      description: "all informations about User"
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/UserID"
        username:
          $ref: "#/components/schemas/Username"
#        follow_list:
#          $ref: "#/components/schemas/Followed"
#        ban_list:
#          $ref: "#/components/schemas/Banned"
#        liked_list:
#          $ref: "#/components/schemas/Liked"
...